requirements.txt
# Core dependencies
python-dotenv>=1.0.0
Pillow>=10.0.0  # For image processing
pytesseract>=0.3.10  # For OCR

# LLM Provider dependencies
google-generativeai>=0.3.0  # For Gemini
openai>=1.0.0  # For OpenAI